1. rename op type to OP_CREATE_CATALOG/OP_DROP_CATALOG 2. handle exception in case of unit test failure 3. change sql stmt package
