+++ fe/fe-core/src/main/java/com/starrocks/persist/EditLog.java
@@ -812,12 +812,12 @@ public class EditLog {
                     globalStateMgr.replayModifyHiveTableColumn(opCode, modifyTableColumnOperationLog);
                     break;
                 }
-                case OperationType.OP_CREATE_EXTERNAL_CATALOG: {
+                case OperationType.OP_CREATE_CATALOG: {
                     CreateCatalogLog createCatalogLog =
                             (CreateCatalogLog) journal.getData();
                     globalStateMgr.getCatalogMgr().replayCreateCatalog(createCatalogLog);
                 }
-                case OperationType.OP_DROP_EXTERNAL_CATALOG: {
+                case OperationType.OP_DROP_CATALOG: {
                     DropCatalogLog dropCatalogLog =
                             (DropCatalogLog) journal.getData();
                     globalStateMgr.getCatalogMgr().replayDropCatalog(dropCatalogLog);
