+++ fe/fe-core/src/main/java/com/starrocks/persist/OperationType.java
@@ -211,9 +211,9 @@ public class OperationType {
     public static final short OP_HEARTBEAT_V2 = 10041; // V2 version of heartbeat
 
     // create external catalog
-    public static final short OP_CREATE_EXTERNAL_CATALOG = 10051;
+    public static final short OP_CREATE_CATALOG = 10051;
 
     // drop external catalog
-    public static final short OP_DROP_EXTERNAL_CATALOG = 10061;
+    public static final short OP_DROP_CATALOG = 10061;
 
 }
